+- Connected signals to <li> elements (fixes links within lists).  - Enabled text and background color-choice in preferences.  - Enabled clicking over image links.    Patches: place +- Fixed a va_list-related SEGFAULT on 64bit-arch in dStr_vsprintfa().    Patch: Vincent Thomasset +- Fixed void to int conversions for 64bit-arch.    Patch: Jorge Arellano, higuita +- Added a strndup() replacement in dw2    Patch: Alexander Becher, Johannes Hofmann, Jorge Arellano +- Fixed calcHashValue() to only return non-negative numbers (was SEGFAULT).  - Improved scrolling performance on large pages by copying screen data     instead of rendering.    Patches: Johannes Hofmann

+++ dlib/dlib.h
@@ -33,7 +33,8 @@ extern "C" {
  *-- Casts -------------------------------------------------------------------
  */
 /* TODO: include a void* size test in configure.in */
-#define VOIDP2INT(p)    ((int)(p))
+/* (long) works for both 32bit and 64bit */
+#define VOIDP2INT(p)    ((long)(p))
 #define INT2VOIDP(i)    ((void*)(i))
 
 /*
